+++ src/config/config.mjs
@@ -1,7 +1,7 @@
 /* eslint-disable space-before-function-paren */
 import { getTestFolderFromArgs, getTestNameFromArgs } from '../cli.mjs'
 import { RunnerError } from '../errors/runner.mjs'
-import { readConfigFile } from '../setup.mjs'
+import { readConfigFile } from './setup.mjs'
 import { applyColor } from '../transform.mjs'
 
 export const configName = 'test.config.js'
