+++ src/core.mjs
@@ -3,7 +3,7 @@ import fs from 'fs'
 import { applyColor, transformStackTrace } from './transform.mjs'
 import { runParsedBlocks, report } from './context.mjs'
 import { getConfig } from './config/config.mjs'
-import { getMultipleFilePath } from './setup.mjs'
+import { getMultipleFilePath } from './config/setup.mjs'
 import { timeStamp, printExecutionTime } from './support.mjs'
 import { createReport } from './reporters/reporter.mjs'
 import { EXIT_CODES } from './constants.mjs'
